requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-jose[cryptography]==3.5.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
